kidlog_project/requirements.txt
asgiref==3.9.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.5
idna==3.11
load-dotenv==0.1.0
pillow==11.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
